Refactor login functionality and enhance UI components; add StartPage and related styles
