+++ frontend/src/components/Favorites.css
@@ -1,0 +1,36 @@
+.favorites-list {
+  display: flex;
+  flex-wrap: wrap;
+  list-style: none;
+}
+
+.favorite-item {
+  margin: 0 0 10px 0;
+  padding: 10px;
+  border: 1px solid #ccc;
+  border-radius: 5px;
+  background-color: #f9f9f9;
+  box-shadow: 0 2px 2px rgba(0, 0, 0, 0.1);
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  gap: 10px;
+}
+
+.favorite-title:hover {
+  cursor: pointer;
+  text-decoration: underline;
+}
+
+/* media queries for responsive design */
+@media (max-width: 768px) {
+  .favorites {
+    margin-top: 2rem;
+  }
+
+  .favorites h2 {
+    margin-bottom: 1rem;
+    text-align: center;
+  }
+}
