+++ frontend/src/pages/Home.css
@@ -2,7 +2,6 @@
   display: flex;
   flex-direction: column;
   gap: 30px;
-  /* justify-content: center; */
   align-items: center;
   min-height: 65vh;
 }
@@ -20,13 +19,7 @@
   gap: 20px;
 }
 
-.btn-suggestion {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-between;
-}
-
+.btn-suggestion,
 .btn-ingredient {
   display: flex;
   flex-direction: row;
@@ -59,9 +52,15 @@
 
 /* media queries for responsive design */
 @media (max-width: 768px) {
+  .home-page {
+    width: 90%;
+    margin: auto;
+  }
+
   .home-content {
     flex-direction: column;
-    margin: 0 2rem;
+    align-items: center;
+    text-align: justify;
   }
 
   .search-section {
